requirements.txt
Bio==1.5.3
joblib==1.1.0
keras==2.11.0
numpy==1.23.5
pandas==1.4.4
Pillow==9.4.0
requests==2.28.1
scikit_learn==1.2.0
scipy==1.10.0
tensorflow==2.11.0
tqdm==4.64.1
